move: AppShell 관련된 파일의 폴더 구조 변경
