+++ src/components/Layout/AppShell.tsx
@@ -1,5 +1,5 @@
-import Footer from '@/components/Footer/Footer';
-import Navbar from '@/components/Navbar/Navbar';
+import Footer from '@/components/Footer';
+import Navbar from '@/components/Navbar';
 
 const AppShell = ({ children }: { children: React.ReactNode }) => {
   return (
